+++ saerojinro-api/src/main/java/goorm/saerojinro/api/reservation/presentation/ReservationControllerImpl.java
@@ -1,6 +1,6 @@
 package goorm.saerojinro.api.reservation.presentation;
 
-import goorm.saerojinro.api.reservation.api.ReservationFacade;
+import goorm.saerojinro.api.reservation.application.ReservationFacade;
 import goorm.saerojinro.api.reservation.presentation.response.ReservationCancelResponse;
 import goorm.saerojinro.api.reservation.presentation.response.ReservationCreateResponse;
 import lombok.RequiredArgsConstructor;
